requirements.txt
python-dotenv
mysql-connector-python
pandas
json
ast
tqdm